test(bm25): fixed term frequency problems
